Move sources to a new module
